+++ client/web/src/enterprise/searchContexts/DeleteSearchContextModal.tsx
@@ -8,9 +8,9 @@ import { LoadingSpinner } from '@sourcegraph/react-loading-spinner'
 import { ISearchContext } from '@sourcegraph/shared/src/graphql/schema'
 import { asError, isErrorLike } from '@sourcegraph/shared/src/util/errors'
 import { useEventObservable } from '@sourcegraph/shared/src/util/useObservable'
+import { ALLOW_NAVIGATION } from '@sourcegraph/web/src/components/AwayPrompt'
 
-import { ALLOW_NAVIGATION } from '../components/AwayPrompt'
-import { SearchContextProps } from '../search'
+import { SearchContextProps } from '../../search'
 
 interface DeleteSearchContextModalProps extends Pick<SearchContextProps, 'deleteSearchContext'> {
     isOpen: boolean