+++ enterprise/cmd/frontend/internal/searchcontexts/resolvers/resolvers_test.go
@@ -1,4 +1,4 @@
-package graphqlbackend
+package resolvers
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 	"github.com/graph-gophers/graphql-go"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
@@ -31,15 +32,15 @@ func TestAutoDefinedSearchContexts(t *testing.T) {
 	database.Mocks.Users.GetByID = func(ctx context.Context, id int32) (*types.User, error) {
 		return &types.User{Username: username}, nil
 	}
-	defer resetMocks()
+	defer func() { database.Mocks.Users.GetByID = nil }()
 
-	searchContexts, err := (&schemaResolver{db: db}).AutoDefinedSearchContexts(ctx)
+	searchContexts, err := (&Resolver{db: db}).AutoDefinedSearchContexts(ctx)
 	if err != nil {
 		t.Fatal(err)
 	}
-	want := []*searchContextResolver{
-		{sc: searchcontexts.GetGlobalSearchContext(), db: db},
-		{sc: searchcontexts.GetUserSearchContext(username, key), db: db},
+	want := []graphqlbackend.SearchContextResolver{
+		&searchContextResolver{sc: searchcontexts.GetGlobalSearchContext(), db: db},
+		&searchContextResolver{sc: searchcontexts.GetUserSearchContext(username, key), db: db},
 	}
 	if !reflect.DeepEqual(searchContexts, want) {
 		t.Fatalf("got %+v, want %+v", searchContexts, want)
@@ -61,28 +62,28 @@ func TestSearchContexts(t *testing.T) {
 	ctx := context.Background()
 
 	userID := int32(1)
-	graphqlUserID := MarshalUserID(userID)
+	graphqlUserID := graphqlbackend.MarshalUserID(userID)
 
 	query := "ctx"
 	tests := []struct {
 		name     string
-		args     *listSearchContextsArgs
+		args     *graphqlbackend.ListSearchContextsArgs
 		wantErr  string
 		wantOpts database.ListSearchContextsOptions
 	}{
 		{
 			name:     "filtering by namespace",
-			args:     &listSearchContextsArgs{Query: &query, Namespaces: []*graphql.ID{&graphqlUserID}},
+			args:     &graphqlbackend.ListSearchContextsArgs{Query: &query, Namespaces: []*graphql.ID{&graphqlUserID}},
 			wantOpts: database.ListSearchContextsOptions{Name: query, NamespaceUserIDs: []int32{userID}, NamespaceOrgIDs: []int32{}, OrderBy: database.SearchContextsOrderBySpec},
 		},
 		{
 			name:     "filtering by instance",
-			args:     &listSearchContextsArgs{Query: &query, Namespaces: []*graphql.ID{nil}},
+			args:     &graphqlbackend.ListSearchContextsArgs{Query: &query, Namespaces: []*graphql.ID{nil}},
 			wantOpts: database.ListSearchContextsOptions{Name: query, NamespaceUserIDs: []int32{}, NamespaceOrgIDs: []int32{}, NoNamespace: true, OrderBy: database.SearchContextsOrderBySpec},
 		},
 		{
 			name:     "get all",
-			args:     &listSearchContextsArgs{Query: &query},
+			args:     &graphqlbackend.ListSearchContextsArgs{Query: &query},
 			wantOpts: database.ListSearchContextsOptions{Name: query, NamespaceUserIDs: []int32{}, NamespaceOrgIDs: []int32{}, OrderBy: database.SearchContextsOrderBySpec},
 		},
 	}
@@ -102,7 +103,7 @@ func TestSearchContexts(t *testing.T) {
 			}
 			defer func() { database.Mocks.SearchContexts.ListSearchContexts = nil }()
 
-			_, err := (&schemaResolver{db: db}).SearchContexts(ctx, tt.args)
+			_, err := (&Resolver{db: db}).SearchContexts(ctx, tt.args)
 			expectErr := tt.wantErr != ""
 			if !expectErr && err != nil {
 				t.Fatalf("expected no error, got %s", err)