+++ internal/gitserver/client.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"crypto/md5"
 	"encoding/binary"
+	"encoding/gob"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -278,6 +279,51 @@ func (c *Cmd) sendExec(ctx context.Context) (_ io.ReadCloser, _ http.Header, err
 		resp.Body.Close()
 		return nil, nil, errors.Errorf("unexpected status code: %d", resp.StatusCode)
 	}
+}
+
+// Search executes a search as specified by args, streaming the results as it goes by calling onMatches with each set of results it
+// receives in response.
+func (c *Client) Search(ctx context.Context, args *protocol.SearchRequest, onMatches func([]protocol.CommitMatch)) (limitHit bool, _ error) {
+	repoName := protocol.NormalizeRepo(args.Repo)
+
+	protocol.RegisterGob()
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	if err := enc.Encode(args); err != nil {
+		return false, err
+	}
+
+	resp, err := c.do(ctx, repoName, "POST", "search", buf.Bytes())
+	if err != nil {
+		return false, err
+	}
+	defer resp.Body.Close()
+
+	var (
+		decodeErr error
+		eventDone protocol.SearchEventDone
+	)
+	dec := StreamSearchDecoder{
+		OnMatches: func(e protocol.SearchEventMatches) {
+			onMatches(e)
+		},
+		OnDone: func(e protocol.SearchEventDone) {
+			eventDone = e
+		},
+		OnUnknown: func(event, _ []byte) {
+			decodeErr = errors.Errorf("unknown event %s", event)
+		},
+	}
+
+	if err := dec.ReadAll(resp.Body); err != nil {
+		return false, err
+	}
+
+	if decodeErr != nil {
+		return false, decodeErr
+	}
+
+	return eventDone.LimitHit, eventDone.Err()
 }
 
 // P4Exec sends a p4 command with given arguments and returns an io.ReadCloser for the output.
@@ -844,12 +890,16 @@ func (c *Client) Remove(ctx context.Context, repo api.RepoName) error {
 }
 
 func (c *Client) httpPost(ctx context.Context, repo api.RepoName, op string, payload interface{}) (resp *http.Response, err error) {
-	return c.do(ctx, repo, "POST", op, payload)
+	b, err := json.Marshal(payload)
+	if err != nil {
+		return nil, err
+	}
+	return c.do(ctx, repo, "POST", op, b)
 }
 
 // do performs a request to a gitserver, sharding based on the given
 // repo name (the repo name is otherwise not used).
-func (c *Client) do(ctx context.Context, repo api.RepoName, method, op string, payload interface{}) (resp *http.Response, err error) {
+func (c *Client) do(ctx context.Context, repo api.RepoName, method, op string, payload []byte) (resp *http.Response, err error) {
 	span, ctx := ot.StartSpanFromContext(ctx, "Client.do")
 	defer func() {
 		span.LogKV("repo", string(repo), "method", method, "op", op)
@@ -860,17 +910,12 @@ func (c *Client) do(ctx context.Context, repo api.RepoName, method, op string, p
 		span.Finish()
 	}()
 
-	reqBody, err := json.Marshal(payload)
-	if err != nil {
-		return nil, err
-	}
-
 	uri := op
 	if !strings.HasPrefix(op, "http") {
 		uri = "http://" + c.AddrForRepo(repo) + "/" + op
 	}
 
-	req, err := http.NewRequest(method, uri, bytes.NewReader(reqBody))
+	req, err := http.NewRequest(method, uri, bytes.NewReader(payload))
 	if err != nil {
 		return nil, err
 	}