+++ migrations/frontend/1528395893_fix_repo_index_part_one.down.sql
@@ -1,0 +1,3 @@
+BEGIN;
+DROP INDEX IF EXISTS repo_non_deleted_id_name_idx;
+COMMIT;